// src/generated/models/Crf82_evaluationsheetsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf82_evaluationsheetsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf82_evaluationsheetsstatecode = keyof typeof Crf82_evaluationsheetsstatecode;
export const Crf82_evaluationsheetsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf82_evaluationsheetsstatuscode = keyof typeof Crf82_evaluationsheetsstatuscode;

export interface Crf82_evaluationsheetsBase {
  crf82_assessmentdate?: string;
  crf82_assessmenttitle?: string;
  crf82_attemptnumber?: number;
  crf82_averagescore?: number;
  crf82_campusname?: string;
  crf82_comments?: string;
  crf82_coursename?: string;
  crf82_coursetype?: string;
  crf82_createdby?: string;
  crf82_creationdate?: string;
  crf82_evaluationsheetid: string;
  crf82_fieldtrainer?: string;
  crf82_generalknowledgescore?: number;
  crf82_instructorassistanttrainer?: string;
  crf82_legacyidentifier?: string;
  crf82_lessonappropriationscore?: number;
  crf82_objectivesjson?: string;
  crf82_passedstatus?: boolean;
  crf82_practicalunderstandingscore?: string;
  crf82_slotidentifier?: string;
  crf82_stepcode?: string;
  crf82_theoreticalunderstandingscore?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf82_evaluationsheetsstatecode;
  statuscode?: Crf82_evaluationsheetsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crf82_evaluationsheets extends Crf82_evaluationsheetsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crf82_passedstatusname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}